Add Order Management in Checkout Feature
Implemented the order management functionality in the checkout feature by adding a new `AddOrderCubit` and its corresponding states. Updated the `get_it_service.dart` to register the `OrdersRepo` and modified the `checkout_view.dart` to use the `AddOrderCubit` for handling order submission. Added a new widget `AddOrderCubitBlocBuilder` to manage the UI state transitions based on the cubit's state.